Move ArchiveFileBlock to format module
